Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Yet, the codec kit is meant to make users able to instantiate custom chains, with possibly filters implemented for things like subtitles.
> 
> Why does this need a codec kit? Subtitles should be some kind of "track"; we already have audio and video data as tracks, so subtitle data should be another track, right?
> 
> Filters should not be part of codecs, but part of the main media node framework, yes? 

No and no. That is the design I agreed with Axel at the moment of the contract. The media nodes should be used only for real time media routing.

> Otherwise we would have two competing processing systems. That would not make much sense to >me.

The codec kit would not replicate the IPC functionalities of media nodes. But provide all the base functionalities that should not be put in realtime IPC.

> 
> > A-la stagefright. I understand you see BMediaFile being simple,
> 
> I don't see it being "simple", I see it being complex but straightforward. I don't see how this is an improvement at all; that's what I've been getting at.
> 
> > The purpose is exactly adding a lower level kit that can support all the stuff that is outlined there and more. If you are not seeing this...I think I can't do a lot to convince you. Sorry : (
> 
> I see how we should indeed support these new things. I do not see how we need or must do it through a new kit, but rather just extend existing APIs, which is what I have been saying. But it seems you do not see what I am trying to explain either...

I see what you are explaining. I already explained there's a lot of stuff out there and other projects had to implement that into advanced frameworks. I encourage you to revert this commit and work your own way through the issues.

Patch-set: 1