Update patch set 1
Patch Set 1: Code-Review-1

As this is a work in progress, it's hard to see the full value of it yet.

Whether it should end up in a new kit or not, can probably be better judged when it's done; before that, it could easily live within the media kit. Since we're not operating in a branch, that part could be reverted, and even should be reverted if R1 is supposed to be created from the master branch, rather than the beta branch.
Since completing this will take time and a lot of changes, and is relatively isolated from other things, this is a good candidate to be pushed forward on a separate branch in github instead.

That aside, I don't mind whether or not BMediaFile will become just a facade, or not -- having a different class as its backend does not mean that class must end up in the public API if there is no good reason for it; BMediaFile could be refactored at a later time to sport the new functionality.

The current API is quite rough, though; I don't like the MouseDown/MouseMoved() hooks at all for example: it's way too specific and insufficient (what about touch screens, keyboard navigation, etc.), and is still different from how the interface kit does things, without any striking (obvious) reason.

In general, I like the idea of modularization and separation; the media node framework doesn't have that much to do with codecs, more with data streams, and buffers. Filters should operate on buffers; they don't really have to be in a codec kit either. It all fits into a media kit, though, but this can be made up from completely different components.

As there are already useful things shaping up (like a DVD add-on), I don't see much reason to completely revert this work, though.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1