Update patch set 1
Patch Set 1:

> 
> The current API is quite rough, though; I don't like the MouseDown/MouseMoved() hooks at all for example: it's way too specific and insufficient

You are right and I planned to introduce something with BHandler, sorry if it wasn't clear in ml.

 (what about touch screens, keyboard navigation, etc.), and is still different from how the interface kit does things, without any striking (obvious) reason.

The whole interface is not touch screen savy so that'd not be a problem. I didn't consider keyboard navigation, you have a very good point there.

> 
> In general, I like the idea of modularization and separation; the media node framework doesn't have that much to do with codecs, more with data streams, and buffers. Filters should operate on buffers; they don't really have to be in a codec kit either. It all fits into a media kit, though, but this can be made up from completely different components.
> 
> As there are already useful things shaping up (like a DVD add-on), I don't see much reason to completely revert this work, though.

You will have to do it yourself however...as in V\OS this kit will continue to be developed as public API and we will get soon to the point where we are incompatible. Too sad, but I can't do much.

Patch-set: 1