Update patch set 1
Patch Set 1:

> > > > If you are saying that BFileInterface is in some fundamental way flawed in its design and not only implementation, and what you are proposing is indeed substantially different and will not have this flaw, of course I want to hear it. But I don't see where you have said anything about this at all.
> > > 
> > > It never worked properly even in BeOS.
> > 
> > Please re-read my first and second paragraphs quoted above: even if it does not work in BeOS, that means not very much about its actual worth, as we have implemented quite a lot of stuff in Haiku that did not work very well in BeOS either.
> 
> We are mixing too much stuff there. BFileInterface is just put in the wrong part. It actually born before BMediaFile when the Be engineers thought they could have done everything with nodes. It turns out it is not possible to do that this way. So it's just a relic.

Why is it not possible to do everything this way? In Windows Media Foundation, an IMFMediaSource is more or less the only way to access codecs, and it is also a node. Their node system supports "pull" as well as "push" model, though. So I don't see how this is "just a relic" when systems designed 10+ years after Be adopted the model successfully, even if Be failed to do it in the 90s.

> > This is not how code review usually works. I have some patches open now (e.g. kernel select > changes) which collected comments that essentially amount to a complete overhaul of the > change (and it's a rather large one.) But I agree it's a much better way of doing things, and > I should clean it up and do that change (it wasn't even mine originally, I adopted it.)
> 
> This is fine, but I'm not willing to do a work that is completely different from my design. If it was just a question of mere suggestions I'd be ok, but that's not the case.

So you are not ever willing to do any work that is not "your design"? That is not how collaborative open source projects work. We debate a design till we find something that we all can agree on, or at least we have left ourselves room to change later. Then someone or multiple people implement it. One can't just declare "my design or I won't work on this", that's not fair, and expect to be a member of the project.

But I see you have sent an email resigning responsibility for the Media Kit, so I guess that is the path you have chosen.

> > > You are not probably going to support streaming servers with BMediaFile alone. There's just a lot of stuff that needs to be supported.
> > 
> > OK. What else have I missed?
> 
> Insisting to put very complex functionality, that often needs entire frameworks to work correctly (live555, icecast) into something relatively stupid like BMediaFile, either you discovered pandora's box or you have a very, very limited view of the issue.

live555 and icecast may be complex, but I've not seen anything they require that will not fit into BMediaFile. You can read and write frames from them, yes? There may be specific time demands, but I don't know what else you need here...

> I'm worried by your 1-min and bang comments, sorry, I try to take some time and not look only at headers to actually discover the stuff.

So do I; I think you underestimate how much time I did spend reading headers while replying. But regardless of that, my comments have or don't have technical merit on their own. It doesn't matter how much time I spent writing them.

The fact is that I have worked with other media APIs before (especially Windows Media Foundation, hence why I am using it as an example so often here) and done some low-level work also, so that I generally understand the constraints and the requirements of building such an API. So a lot of this I just know off the top of my head, I don't need to go look it up, and so I am only impeded by how fast I can type.

Patch-set: 1