Update patch set 1
Patch Set 1:

> As this is a work in progress, it's hard to see the full value of it yet.

Indeed; my point is that whatever value it *should* have, i.e. why so much effort is being poured into it, should already be made clear by whoever is spending such effort. It has not been.

> That aside, I don't mind whether or not BMediaFile will become just a facade, or not -- having a different class as its backend does not mean that class must end up in the public API if there is no good reason for it; BMediaFile could be refactored at a later time to sport the new functionality.

Again, as I have been explaining in this thread, I cannot think of any functionality which could not be added to BMediaFile cleanly, and Barrett has not explained much of anything either.

We already have a "different class" as the BMediaFile backend, though: the plugin system (FFmpeg, etc. plugins). This is a "lower-level API" indeed, but this is because BMediaFile and the rest of the media kit takes care of a variety of other things that the codecs do not need to, but must be done for reading and writing files. So the distinction makes sense.

> In general, I like the idea of modularization and separation; the media node framework doesn't have that much to do with codecs, more with data streams, and buffers. Filters should operate on buffers; they don't really have to be in a codec kit either. It all fits into a media kit, though, but this can be made up from completely different components.

In order to do media processing, one needs a system of interconnecting components: so, a "media source", and then "processors" (filters, compositors, etc.), and then a "sink" (file, screen, etc.) Be mixed a lot of metaphors in their Media Kit and it did not turn out very well. Future APIs, like Windows Media Foundation, mixed most of the same metaphors and it turned out very well. I think a lot of the difference here is the ability to run in "pull" mode rather than "push" mode, which solves a lot of problems here; but this is a different issue.

Having two different node systems is absolutely a bad idea, though.

> As there are already useful things shaping up (like a DVD add-on), I don't see much reason to completely revert this work, though.

The DVD add-on could be done in large part via FFmpeg, I believe; I'm not sure we need to make a new add-on for it. I haven't ever implemented a DVD player before, so I could be mistaken on that point at least.

But the DVD plugin could also exist in the old API for the most part. I'll try to preserve it in the next iteration of this commit.

Patch-set: 1