Update patch set 1
Patch Set 1:

> Why is it not possible to do everything this way? In Windows Media Foundation, an IMFMediaSource is more or less the only way to access codecs, and it is also a node. Their node system supports "pull" as well as "push" model, though. So I don't see how this is "just a relic" when systems designed 10+ years after Be adopted the model successfully, even if Be failed to do it in the 90s.

It's not possible with the current media kit. But media2 would work in a similar way. My draft also includes work on supporting both push and pull API. Unfortunately, it will never be discussed in the area of this OS.

> 
> > > This is not how code review usually works. I have some patches open now (e.g. kernel select > changes) which collected comments that essentially amount to a complete overhaul of the > change (and it's a rather large one.) But I agree it's a much better way of doing things, and > I should clean it up and do that change (it wasn't even mine originally, I adopted it.)
> > 
> > This is fine, but I'm not willing to do a work that is completely different from my design. If it was just a question of mere suggestions I'd be ok, but that's not the case.
> 
> So you are not ever willing to do any work that is not "your design"? That is not how collaborative open source projects work. We debate a design till we find something that we all can agree on, or at least we have left ourselves room to change later. Then someone or multiple people implement it. One can't just declare "my design or I won't work on this", that's not fair, and expect to be a member of the project.

If I fail to convince people with my design I let them enjoy the work. If the design that comes out is wrong for me, I'm not willing to do the work. I think it's perfectly fine leaving the area and let others work on the issue. That's what happened with the buildbots for instance. I think we could agree on a common design, and I think your opinion could be valuable too. It's just that at this moment I'm not willing to do that for a variety of reasons.

Patch-set: 1