Update patch set 1
Patch Set 1: Code-Review+1

I think someone should put a branch on GitHub with the current state of this code, then we should apply this change (minus of course the few areas waddlesplash noticed.) Of course regardless we will have the Git history so maybe the branch isn't needed. But it makes it a bit easier to look at this work later if we need it.

For whatever reason, Dario does not want to take any of our ideas or suggestions. It is his design, or nothing. I think that is pretty unreasonable. These changes to create the Codec Kit are too extensive with no obvious benefit. I don't think this was properly reviewed before, or, if it was, I bet it devolved into a similar argument and everyone else got fed up. I assume this is where Adrien got upset.

I will try to take on some Media Kit work in the next few weeks.

Patch-set: 1
Reviewer: Gerrit User 1000109 <1000109@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1