Update patch set 1
Patch Set 1:

> Patch Set 1: 
> 
> If the suggestions are just "revert it" I don't think there's compromise. And I'm also OK with accepting that I did something wrong, I mean, I'm not going to argue against reverting it if you think it's not fitting. I can't see a more reasonable behavior than that.

You started the revert process, so don't act like we put you in this position. It is not about you doing anything wrong, just about some of us evaluating the Codec Kit changes after the fact. I was not involved in Haiku development for many years, as you may know. So I did not have a chance to review these changes before. I also did not know the Media Kit code, and I am still learning it.

These changes seem to be too much for what they claim to fix. Maybe over time we will learn you were right in aspects. If so, I certainly will give you credit. For now though, it seems the safest route for Haiku is to go back to how the code was before, to more of a BeOS design. If that makes our code incompatible with V\OS, unfortunately I think it will hurt V\OS more than Haiku. We aren't doing this to be mean to you, we just want what is best for Haiku and its development at the moment.

Patch-set: 1