Update patch set 1
Patch Set 1:

> If I fail to convince people with my design I let them enjoy the work. If the design that comes out is wrong for me, I'm not willing to do the work. I think it's perfectly fine leaving the area and let others work on the issue. That's what happened with the buildbots for instance. I think we could agree on a common design, and I think your opinion could be valuable too. It's just that at this moment I'm not willing to do that for a variety of reasons.

No one can force you to work on something you don't want to; I guess we're all on the same page there. However, you didn't fail to convince people with your design -- you didn't really try yet; what's missing is a piece that actually explains your vision in more or less detail. If you have written up something like that before, a link would be fine (to a single document, not a thread of 20 separate mails).

It's hard to judge your work without knowing the full picture; the BMediaFile vs. whatever discussion is pretty irrelevant in the end. What matters is the functionality the new API layer is supposed to bring. If that whole stuff ends up in a media kit, or a codec kit, or somewhere else, cannot really be decided upfront, especially when the whole picture isn't really known.

Patch-set: 1