Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > So having two systems (media node, codec APIs) that do the same thing (filters, effects, etc.) but in two different contexts is frankly ridiculous. I see no reason why we would ever do that, and I see plenty of other systems that did not do that and they have turned out just fine.
> 
> I don't think it's the same thing: I can apply filters to a single image. Why would I want to setup a whole media node graph just for that?

You brought up a point that I wasn't able to recap. This is along another 2-3 points the major media kit flaw. The fact that (with a BMediaEventLooper) you have to spawn a thread per node. This is not going to work outside the simple cases. So, my media2 is designed to take this into account and the codec kit is designed with this in mind too.

> Functionality that can be separated, and may even be useful separately, should be separated IMO; in the API only if the latter proves to be true. For example, if you'd do a non-destructive image processor, wouldn't it be nice if you could easily apply high quality system delivered filters?
> 
> Besides that, the API is always the deciding factor; if that one is sane, and I can get simple things done without too much overhead and boilerplate code, I don't care if I'm using media nodes behind the scenes. I just don't want them in my face in that case :-)

Using media nodes in this case is going to be a complete mess. Not only for this reason, also because all the B_OFFLINE stuff is severely broken and will lead to a lot of overhead (why I need to use ports locally for example?).

With my API the idea is that you instantiate a BMediaStreamer and fed it into a BMediaExtractor. Then when you read it, you are able to select the filter (in this case a decoder), for example you may want to use ASS subtitles or bitmaps those are really different situations. When media2 was ready, I'd have made it a media source, that would have allowed to see each track as an endpoint, and convert from pull to push API. In general, when reading codecs is much more convenient a pull API, while for realtime stuff a push API is way better.

You can do that in a few lines of code, you can select which streamer for use and in some cases you may not want to use a streamer at all, and fed into the extractor just a BDataIO.

Patch-set: 1