Update patch set 1
Patch Set 1:

> > > So having two systems (media node, codec APIs) that do the same thing (filters, effects, etc.) but in two different contexts is frankly ridiculous. I see no reason why we would ever do that, and I see plenty of other systems that did not do that and they have turned out just fine.
> > 
> > I don't think it's the same thing: I can apply filters to a single image. Why would I want to setup a whole media node graph just for that?

The media system is designed to process video or audio as fast as computationally possible. As a result there is a lot of overhead for setup that would not be required for a single image. So what you are talking about here sounds more like the translators, not the media framework.

One could define an API on top of the media graph API to do this though; in fact GIMP's new API internally is rather like a media pipeline in this respect. But the Codec Kit wouldn't solve this problem; it just adds more boilerplate, so that question is ultimately not relevant right now.

> You brought up a point that I wasn't able to recap. This is along another 2-3 points the major media kit flaw. The fact that (with a BMediaEventLooper) you have to spawn a thread per node. This is not going to work outside the simple cases. So, my media2 is designed to take this into account and the codec kit is designed with this in mind too.

BMediaFile/BMediaTrack do not care what thread they are used from, only that they are used from one thread at a time. So the BMediaNode threading problems do not demand a Codec Kit; and if you are saying that we need a new media node API because of them, that may or may not be the case, but at any rate it does not mean we need a new graph API separate from the media graph API, even if the media graph API needs a redesign.

I also am not sure why it is a problem that one has a "thread per node." Actually I don't see why you need a thread per node at all; a BMediaEventLooper is *derived from* a BMediaNode; the fact that most BMediaNodes are also BMediaEventLoopers is a different problem altogether. But I am really not discussing the merits and demerits of the existing Media Kit overall, only the media file related classes here; this is a problem for another time.

> Using media nodes in this case is going to be a complete mess. Not only for this reason, also because all the B_OFFLINE stuff is severely broken and will lead to a lot of overhead (why I need to use ports locally for example?).

All BLoopers, and thus every window in Haiku, use a port. Even if nothing will talk to them besides other threads in that application, they use a port. So I don't see how this question makes much sense; are you saying the Interface Kit "one window, one thread" design is a bad one, too?

Again, B_OFFLINE stuff being broken or having overhead does *not inherently mean* that the system cannot ever be that way. Unless you have a reason or an argument why the system is fundamentally flawed and that functionality can never work, simply stating it just means it does not work *right now*.

> In general, when reading codecs is much more convenient a pull API, while for realtime stuff a push API is way better.

I agree with this.

It seems your methodology about how to deal with this fact I completely disagree with though, as is evidenced by our disagreements in this thread...

> You can do that in a few lines of code, you can select which streamer for use and in some cases you may not want to use a streamer at all, and fed into the extractor just a BDataIO.

And again, my entire argument here was that BAdapterIO is more than enough for media streaming, one does not need a BMediaStreamer as a separate class, particularly as the BAdapterIO could be used for other things (e.g. a remote SQlite file could be accessed via a BAdapterIO+HTTP"Streamer".) Why restrict this so much?

Patch-set: 1