Update patch set 1
Patch Set 1:

> I also am not sure why it is a problem that one has a "thread per node." Actually I don't see why you need a thread per node at all; a BMediaEventLooper is *derived from* a BMediaNode; the fact that most BMediaNodes are also BMediaEventLoopers is a different problem altogether. But I am really not discussing the merits and demerits of the existing Media Kit overall, only the media file related classes here; this is a problem for another time.

Let's suppose we want to decode an a/v file, filter it and reproduce, here's what will happen using a pure media_kit approach:

* 1 thread read the file (BFileInterface)
* 1 thread demux the raw data
* 2 threads filter the audio and video
* (video) 1 thread receive the data and process it
* (video) 1 thread (the window) draw it on destination
* (audio) 1 thread (the mixer) mix it
* (audio) 1 thread (the multiaudio) get the final data and play it

In the whole chain we have 9 (!?) threads, without even considering all the locking and context switch in the middle it's going to be a pity. The context switch happens because you are probably going to host the stuff in another process by making those modules available in the system graph. This is supposing that there's no more routing, this is just the simplest example ever.

Let's take for example using the codec kit:

* 1 Thread read and demux the data and apply the filters
* (video) 1 thread read and write to window (either a media node or custom stuff don't matter)
* (audio) 1 thread send the data to the mixer
* (audio) 1 thread get the data to the soundcard (multiaudio)

We have "only" 4 threads. If you are telling me the first chain is more efficient, I'm done.

Patch-set: 1