Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > > > I don't think it's the same thing: I can apply filters to a single image. Why would I want to setup a whole media node graph just for that?
> > It seems you are not getting the point. There is no graph in the codec kit.
> 
> You obviously missed that was indeed why I said that... I don't need a media node for that, that's why that functionality should not only be accessible behind that framework. But it makes total sense to me to have a media node that can apply filters. Filters are just not media node specific but something more generic.

That's why I think gerrit is not proper to discuss this stuff. Honestly, it is getting me headaches.

I quoted that line like I've been replying to waddlesplash and read it outside of the context. In fact I even replied directly to your comment saying that it's a good point...

> 
> > > One could define an API on top of the media graph API to do this though; in fact GIMP's new API internally is rather like a media pipeline in this respect. But the Codec Kit wouldn't solve this problem; it just adds more boilerplate, so that question is ultimately not relevant right now.
> > Do it yourself, I already know enough to say that it's going to fail.
> 
> Not a helpful comment. If you think it's going to fail (or even know), explain yourself. I see absolutely no reason why this could possibly fail. It would just not be the way I'd implement it. I also can't see how the codec kit would add more boilerplate -- that's entirely up to how its API will be designed.

Yes "fail" is not the appropriate term. I meant that it'd be very very suboptimal and I explained that already...

> 
> > Try to work with media nodes without a centralized event loop and you will realize how nice it is...
> 
> That I don't understand. There is no centralized event loop in the media nodes. There is one per node. 

That's what I said...not sure why you implied I meant a "central loop for all media nodes in the system". I've been referring of course to BMediaEventLooper like in the earlier email...

> Are you suggesting a framework where we can directly call methods of other nodes, and there are "bridge" nodes for when the team boundary is crossed? 

This is how I'm planning media2 to work...

> Is that what you mean by that? Because I guess everyone would understand the benefit of that (compared to an event loop).

hopefully :)

> 
> > It's a completely different situation I'd not compare them at all.
> 
> Yes, and no. That media nodes use ports even internally is an implementation detail. It's nothing that needs to be done this way. 

I'm not sure what you mean there, as I read it it's wrong. BMediaNode knows indeed about ports. In fact it's the default way you should contact it. Maybe it can be removed, but that'd be a problem too big to be justifiable.

>The application kit actually already implements a direct messenger for this use case: communication within a team does not use ports at all.
> Media nodes could do exactly the same thing.

But with nodes it'd be much different because BMediaNode actually export it's port id and it's supposed that HandleMessage is called. It'd be a bit more difficult as I understand it, because I don't think messengers export the port.

> 
> > No there is serious overhead and unneeded complexity in the code if you want this to fit well with other kind of nodes. It could even work, but in the end you will have to use hundreds of lines of code for each stupid thing you want to do.
> 
> Of course it could work, but I think we can all agree that the media node API is too complex to use, and really needs a rework.

Yeah, that's one of the points. It's actually waddlesplash that want to do like Be devs did in BeOS R4, not me :)

Patch-set: 1