Update patch set 1
Patch Set 1:

> > If there will be a "filter" system, there will be a graph in the codec kit, even if only a flat one.
> 
> No. A graph is something much more complex than a simple linear chain of components. Also, the codec kit chain is not meant to be IPC.

A linear chain of interchangeable components is a graph node framework. It is a linear graph, but a graph nonetheless. Nowhere is it said this must be IPC-able.

> > > No there is serious overhead and unneeded complexity in the code if you want this to fit well with other kind of nodes. It could even work, but in the end you will have to use hundreds of lines of code for each stupid thing you want to do.
> > 
> > That does not mean it can't work. It just means we need a higher-level API for it.
> 
> Okay, waiting to see your API.

Believe me, I'd like to design one at this point if only to show it can be done and it can be efficient and easy to program. But I'm too busy with XHCI, GPU acceleration, R1 bugfixing, ...

> > > > And again, my entire argument here was that BAdapterIO is more than enough for media streaming, one does not need a BMediaStreamer as a separate class, particularly as the BAdapterIO could be used for other things (e.g. a remote SQlite file could be accessed via a BAdapterIO+HTTP"Streamer".) Why restrict this so much?
> > > 
> > > That doesn't make sense, sorry.
> > 
> > Again, why not?
> 
> As I already explained, the BAdapterIO is just a low level handle to read and write data. It knows basically nothing about what's happening on the upper level, and it's designed this way because streamers may want to do more and they have to do it in the BStreamer.

What "more" will streamers want to do that a BAdapterIO cannot? You still have not explained this; as far as I can tell, the things you have mentioned are all in the realm of filters and decoders, not the I/O portion.

Patch-set: 1