Update patch set 1
Patch Set 1:

> That being said, I don't like to use the "others do it this way" argument in general. But the libstagefright API works more or less like the codec kit. It's possible to chain components, select them and so on. So if you want use case I suggest you to look at how many things are implemented using this API (OpenMAX, streaming clients and servers, handling images like translators) in Android.

libstagefright is a media playback engine that also includes a minimal graph node system for processing resultant data. For what it is, it's fine; but for us, it's not because we don't need a graph node system in both a "Codec" kit and a "Media" kit. If the graph node system of the Media kit is unsuitable, then we need to design a new one that is for all purposes, not just some.

Patch-set: 1