Update patch set 1
Patch Set 1:

> In the whole chain we have 9 (!?) threads, without even considering all the locking and context switch in the middle it's going to be a pity.

Haiku is designed around pervasive multi-threading. Locking as long as there is little contention is cheap; and context switching is also cheap. So this is a problem because of the complexity of setting it up, but not so because of the locking or context switches, necessarily. Doing away with those would be nice but not required.

> The context switch happens because you are probably going to host the stuff in another process by making those modules available in the system graph. This is supposing that there's no more routing, this is just the simplest example ever.

I am in full agreement, as I have already said, that the media kit needs to be seriously rethought. I do not agree the Codec Kit is in the right direction at all.

> Let's take for example using the codec kit:
> 
> * 1 Thread read and demux the data and apply the filters
> * (video) 1 thread read and write to window (either a media node or custom stuff don't matter)
> * (audio) 1 thread send the data to the mixer
> * (audio) 1 thread get the data to the soundcard (multiaudio)
> 
> We have "only" 4 threads. If you are telling me the first chain is more efficient, I'm done.

Of course your chain is more efficient. I am not claiming it isn't; I am claiming that you are creating a new graph node system that will compete with the Media Kit one. I already said that we should probably invent a Media Kit 2 with a better graph node system; but inventing a Codec Kit to make reading files "simpler" when in actuality you are creating a lot of complexity under the hood is kind of ridiculous.

Patch-set: 1