Update patch set 1
Patch Set 1:

> > A linear chain of interchangeable components is a graph node framework. It is a linear graph, but a graph nonetheless. Nowhere is it said this must be IPC-able.
> 
> :) then everything is a graph. You are a graph. The universe is a graph. Come on, let's keep it technical. Technically it's not a graph but a stack.

I am "a chain of interchangeable components"? So is "the universe"? Well obviously that is false.

No, it's a graph-node framework. You are going to process audio and video differently, yes? So then it won't even be a linear graph, it will have a fork at the beginning after the demux occurs, and then two different sinks (a screen and a speaker.) So thus you have an actual tree, and thus it is a graph.

> > Believe me, I'd like to design one at this point if only to show it can be done and it can be efficient and easy to program. But I'm too busy with XHCI, GPU acceleration, R1 bugfixing,
> 
> I'm sad you don't admit how the current solution is suboptimal.

I already admitted it is. However, it worked for BeOS on 144MHz machines in the 1990s; so it seems its sub-optimal-ness is mostly in how hard it is to program, and then also in the latency problems. We can live with those for now, and design MediaKit2 after R1.

But even if R1 were out, I would still be voting against Codec Kit, because it is a worse solution than the current sub-optimal-ness.
 
> > What "more" will streamers want to do that a BAdapterIO cannot? You still have not explained this; as far as I can tell, the things you have mentioned are all in the realm of filters and decoders, not the I/O portion.
> 
> In theory every class can do everything. The BStreamer is actually responsible to do the sniffing. The BAdapterIO is the i/o handle. At the time of the contract that's the solution that came more reasonable by splitting functionalities. That's how it is designed, you have to show me why I should merge them...not the opposite.

I am not saying to merge them. BAdapterIO is and should remain the I/O handle. I don't think there should be a BStreamer though; the existing plugin API already did OK with sniffing for the most part. So BStreamer would be entirely superfluous as far as I can see.

Patch-set: 1